+++ benches/bench.rs
@@ -2,8 +2,7 @@
 
 extern crate test;
 
-use divbuf::{DivBufMut, DivBufShared};
-use futures::sync::oneshot;
+use futures::channel::oneshot;
 use std::fs;
 use std::io::Write;
 use std::os::unix::fs::FileExt;
@@ -11,7 +10,7 @@ use std::sync::{Arc, Mutex, mpsc};
 use std::thread;
 use tempfile::TempDir;
 use test::Bencher;
-use tokio::runtime::current_thread::Runtime;
+use tokio::runtime::Runtime;
 use tokio_file::File;
 
 const FLEN: usize = 1<<19;
@@ -23,20 +22,18 @@ fn bench_aio_read(bench: &mut Bencher) {
     let path = dir.path().join("aio_read");
     let mut f = fs::File::create(&path).unwrap();
     let wbuf = vec![0; FLEN];
-    let dbs = DivBufShared::from(vec![0; FLEN]);
     f.write_all(&wbuf).expect("write failed");
 
     // Prep the reactor
     let mut runtime = Runtime::new().unwrap();
     let file = File::open(path).unwrap();
 
+    let mut rbuf = vec![0u8; FLEN];
     bench.iter(move || {
-        let rbuf = Box::new(dbs.try_mut().unwrap());
-        let fut = file.read_at(rbuf, 0)
-            .expect("read_at failed early");
-        let len = runtime.block_on(fut)
-                         .unwrap()
-                         .value.unwrap();
+        let len = runtime.block_on(async{
+            file.read_at(&mut rbuf[..], 0)
+                .expect("read_at failed early").await
+        }).unwrap().value.unwrap();
         assert_eq!(len as usize, wbuf.len());
     })
 }
@@ -50,7 +47,6 @@ fn bench_threaded_read(bench: &mut Bencher) {
     let path = dir.path().join("threaded_read");
     let mut f = fs::File::create(&path).unwrap();
     let wbuf = vec![0; FLEN];
-    let dbs = DivBufShared::from(vec![0; FLEN]);
     f.write_all(&wbuf).expect("write failed");
 
     // Prep the reactor
@@ -59,14 +55,13 @@ fn bench_threaded_read(bench: &mut Bencher) {
 
     bench.iter(move || {
         let (tx, rx) = oneshot::channel::<usize>();
-        let mut rbuf = dbs.try_mut().unwrap();
         let fclone = file.clone();
         thread::spawn(move || {
+            let mut rbuf = vec![0u8; FLEN];
             let len = fclone.lock()
                             .unwrap()
                             .read_at(&mut rbuf[..], 0)
                             .expect("read failed");
-            drop(rbuf);
             tx.send(len).expect("sending failed");
         });
         let len = runtime.block_on(rx).expect("receiving failed");
@@ -76,7 +71,6 @@ fn bench_threaded_read(bench: &mut Bencher) {
     
 struct TpOpspec {
     f:  Arc<Mutex<fs::File>>,
-    buf: DivBufMut,
     offset: u64,
     tx: oneshot::Sender<usize>
 }
@@ -98,13 +92,12 @@ fn bench_threadpool_read(bench: &mut Bencher) {
     thread::spawn(move || {
         loop {
             let v: Option<TpOpspec> = prx.recv().unwrap();
-            if let Some(mut op) = v {
+            if let Some(op) = v {
+                let mut rbuf = vec![0; FLEN];
                 let f = op.f.lock().unwrap();
                 let len = {
-                    let sl = &mut op.buf[..];
-                    f.read_at(sl, op.offset).unwrap()
+                    f.read_at(&mut rbuf[..], op.offset).unwrap()
                 };
-                drop(op.buf);
                 op.tx.send(len).expect("send failed");
             } else {
                 break;
@@ -118,11 +111,9 @@ fn bench_threadpool_read(bench: &mut Bencher) {
     let ptxclone = ptx.clone();
 
     bench.iter(move || {
-        let dbs = DivBufShared::from(vec![0; FLEN]);
         let (tx, rx) = oneshot::channel::<usize>();
         let opspec = TpOpspec {
             f: file.clone(),
-            buf: dbs.try_mut().unwrap(),
             offset: 0,
             tx
         };